+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/TransferManager.cs
@@ -10,6 +10,7 @@ using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
 using Azure.Core;
+using Azure.Core.Pipeline;
 using Azure.Storage.DataMovement.Models;
 using Azure.Storage.DataMovement.Models.JobPlan;
 
@@ -71,11 +72,11 @@ namespace Azure.Storage.DataMovement
 
         /// <summary>
         /// Defines the error handling method to follow when an error is seen. Defaults to
-        /// <see cref="ErrorHandlingOptions.StopOnAllFailures"/>.
+        /// <see cref="ErrorHandlingBehavior.StopOnAllFailures"/>.
         ///
-        /// See <see cref="ErrorHandlingOptions"/>.
+        /// See <see cref="ErrorHandlingBehavior"/>.
         /// </summary>
-        internal ErrorHandlingOptions _errorHandling;
+        internal ErrorHandlingBehavior _errorHandling;
 
         /// <summary>
         /// Cancels the channels operations when disposing.
@@ -88,6 +89,8 @@ namespace Azure.Storage.DataMovement
         /// </summary>
         internal ArrayPool<byte> UploadArrayPool => _arrayPool;
         private ArrayPool<byte> _arrayPool;
+
+        internal ClientDiagnostics ClientDiagnostics { get; }
 
         /// <summary>
         /// Protected constructor for mocking.
@@ -127,7 +130,8 @@ namespace Azure.Storage.DataMovement
             _checkpointer = _checkpointerOptions != default ? _checkpointerOptions.GetCheckpointer() : CreateDefaultCheckpointer();
             _dataTransfers = new Dictionary<string, DataTransfer>();
             _arrayPool = ArrayPool<byte>.Shared;
-            _errorHandling = options?.ErrorHandling != default ? options.ErrorHandling : ErrorHandlingOptions.StopOnAllFailures;
+            _errorHandling = options?.ErrorHandling != default ? options.ErrorHandling : ErrorHandlingBehavior.StopOnAllFailures;
+            ClientDiagnostics = new ClientDiagnostics(options?.ClientOptions ?? ClientOptions.Default);
         }
 
         #region Job Channel Management
@@ -361,6 +365,11 @@ namespace Azure.Storage.DataMovement
             CancellationToken cancellationToken = default)
         {
             CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
+            Argument.AssertNotNullOrWhiteSpace(transferId, nameof(transferId));
+            Argument.AssertNotNull(sourceResource, nameof(sourceResource));
+            Argument.AssertNotNull(destinationResource, nameof(destinationResource));
+
+            transferOptions ??= new TransferOptions();
 
             if (_dataTransfers.ContainsKey(transferId))
             {
@@ -429,33 +438,15 @@ namespace Azure.Storage.DataMovement
 
             transferOptions ??= new TransferOptions();
 
-            string transferId;
-            bool resumeJob = false;
-            if (!string.IsNullOrEmpty(transferOptions.ResumeFromCheckpointId))
-            {
-                resumeJob = true;
-                transferId = transferOptions.ResumeFromCheckpointId;
-
-                if (_dataTransfers.ContainsKey(transferId))
-                {
-                    // Remove the stale DataTransfer so we can pass a new DataTransfer object
-                    // to the user and also track the transfer from the DataTransfer object
-                    _dataTransfers.Remove(transferId);
-                }
-            }
-            else
-            {
-                // Add Transfer to Checkpointer
-                transferId = Guid.NewGuid().ToString();
-                await _checkpointer.AddNewJobAsync(transferId, _cancellationToken).ConfigureAwait(false);
-            }
+            string transferId = Guid.NewGuid().ToString();
+            await _checkpointer.AddNewJobAsync(transferId, _cancellationToken).ConfigureAwait(false);
 
             DataTransfer dataTransfer = await BuildAndAddTransferJobAsync(
                 sourceResource,
                 destinationResource,
                 transferOptions,
                 transferId,
-                resumeJob,
+                false,
                 cancellationToken).ConfigureAwait(false);
 
             return dataTransfer;
@@ -516,9 +507,9 @@ namespace Azure.Storage.DataMovement
         {
             // If the resource cannot produce a Uri, it means it can only produce a local path
             // From here we only support an upload job
-            if (sourceResource.CanProduceUri == ProduceUriType.NoUri)
+            if (!sourceResource.CanProduceUri)
             {
-                if (destinationResource.CanProduceUri == ProduceUriType.ProducesUri)
+                if (destinationResource.CanProduceUri)
                 {
                     // Stream to Uri job (Upload Job)
                     StreamToUriTransferJob streamToUriJob = new StreamToUriTransferJob(
@@ -529,7 +520,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -557,7 +549,7 @@ namespace Azure.Storage.DataMovement
             else
             {
                 // Source is remote
-                if (destinationResource.CanProduceUri == ProduceUriType.ProducesUri)
+                if (destinationResource.CanProduceUri)
                 {
                     // Service to Service Job (Copy job)
                     ServiceToServiceTransferJob serviceToServiceJob = new ServiceToServiceTransferJob(
@@ -568,7 +560,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         CheckPointFolderPath: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -600,7 +593,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -632,9 +626,9 @@ namespace Azure.Storage.DataMovement
         {
             // If the resource cannot produce a Uri, it means it can only produce a local path
             // From here we only support an upload job
-            if (sourceResource.CanProduceUri == ProduceUriType.NoUri)
+            if (!sourceResource.CanProduceUri)
             {
-                if (destinationResource.CanProduceUri == ProduceUriType.ProducesUri)
+                if (destinationResource.CanProduceUri)
                 {
                     // Stream to Uri job (Upload Job)
                     StreamToUriTransferJob streamToUriJob = new StreamToUriTransferJob(
@@ -645,7 +639,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -680,7 +675,7 @@ namespace Azure.Storage.DataMovement
             else
             {
                 // Source is remote
-                if (destinationResource.CanProduceUri == ProduceUriType.ProducesUri)
+                if (destinationResource.CanProduceUri)
                 {
                     // Service to Service Job (Copy job)
                     ServiceToServiceTransferJob serviceToServiceJob = new ServiceToServiceTransferJob(
@@ -691,7 +686,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -730,7 +726,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {